pyproject.toml
[project]
name = "predicates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.5",
    "pydash>=8.0.5",
    "pyyaml>=6.0.2",
]
